Consolidate model package
It is better to keep the `MethodHandler` and declaration-related classes
together in the same package. Hopefully, some of the classes can be then
hidden from the public API. This is a preparation step for implementing
code-generated treating of handlers (instead of reflection-based).
