+++ server/src/main/java/io/spine/server/aggregate/model/Applier.java
@@ -26,8 +26,8 @@ import io.spine.base.EventMessage;
 import io.spine.server.aggregate.Aggregate;
 import io.spine.server.model.AbstractHandlerMethod;
 import io.spine.server.model.MethodAttribute;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.model.VoidMethod;
-import io.spine.server.model.declare.ParameterSpec;
 import io.spine.server.type.EmptyClass;
 import io.spine.server.type.EventClass;
 import io.spine.server.type.EventEnvelope;
