+++ server/src/main/java/io/spine/server/command/model/CommandAcceptingMethod.java
@@ -27,7 +27,7 @@ import io.spine.server.EventProducer;
 import io.spine.server.dispatch.Success;
 import io.spine.server.event.RejectionEnvelope;
 import io.spine.server.model.AbstractHandlerMethod;
-import io.spine.server.model.declare.ParameterSpec;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.type.CommandClass;
 import io.spine.server.type.CommandEnvelope;
 import io.spine.type.MessageClass;
