+++ server/src/main/java/io/spine/server/command/model/CommandAcceptingMethodSignature.java
@@ -26,15 +26,15 @@ import io.spine.base.CommandMessage;
 import io.spine.base.ThrowableMessage;
 import io.spine.core.CommandContext;
 import io.spine.server.model.HandlerMethod;
-import io.spine.server.model.declare.MethodSignature;
-import io.spine.server.model.declare.ParameterSpec;
+import io.spine.server.model.MethodSignature;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.type.CommandClass;
 import io.spine.server.type.CommandEnvelope;
 
 import java.lang.annotation.Annotation;
 
-import static io.spine.server.model.declare.MethodParams.consistsOfSingle;
-import static io.spine.server.model.declare.MethodParams.consistsOfTwo;
+import static io.spine.server.model.MethodParams.consistsOfSingle;
+import static io.spine.server.model.MethodParams.consistsOfTwo;
 
 /**
  * The signature of a method, that accepts {@code Command} envelopes as parameter values.
