+++ server/src/main/java/io/spine/server/command/model/CommandHandlerMethod.java
@@ -24,7 +24,7 @@ import io.spine.server.command.CommandHandler;
 import io.spine.server.dispatch.Success;
 import io.spine.server.model.EventProducingMethod;
 import io.spine.server.model.IllegalOutcomeException;
-import io.spine.server.model.declare.ParameterSpec;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.type.CommandClass;
 import io.spine.server.type.CommandEnvelope;
 import io.spine.server.type.EventClass;
