+++ server/src/main/java/io/spine/server/command/model/CommandHandlerSignature.java
@@ -23,7 +23,7 @@ package io.spine.server.command.model;
 import com.google.common.collect.ImmutableSet;
 import io.spine.base.EventMessage;
 import io.spine.server.command.Assign;
-import io.spine.server.model.declare.ParameterSpec;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.type.CommandEnvelope;
 
 import java.lang.reflect.Method;
