+++ server/src/main/java/io/spine/server/command/model/CommandReactionSignature.java
@@ -28,16 +28,16 @@ import io.spine.base.RejectionMessage;
 import io.spine.core.CommandContext;
 import io.spine.core.EventContext;
 import io.spine.server.command.Command;
-import io.spine.server.model.declare.MethodParams;
-import io.spine.server.model.declare.MethodSignature;
-import io.spine.server.model.declare.ParameterSpec;
+import io.spine.server.model.MethodParams;
+import io.spine.server.model.MethodSignature;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.type.EventEnvelope;
 
 import java.lang.reflect.Method;
 import java.util.Optional;
 
-import static io.spine.server.model.declare.MethodParams.consistsOfSingle;
-import static io.spine.server.model.declare.MethodParams.consistsOfTwo;
+import static io.spine.server.model.MethodParams.consistsOfSingle;
+import static io.spine.server.model.MethodParams.consistsOfTwo;
 
 /**
  * A signature of {@link CommandReactionMethod}.
