+++ server/src/main/java/io/spine/server/command/model/CommandSubstituteSignature.java
@@ -23,8 +23,8 @@ package io.spine.server.command.model;
 import com.google.common.collect.ImmutableSet;
 import io.spine.base.CommandMessage;
 import io.spine.server.command.Command;
-import io.spine.server.model.declare.MethodParams;
-import io.spine.server.model.declare.ParameterSpec;
+import io.spine.server.model.MethodParams;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.type.CommandEnvelope;
 
 import java.lang.reflect.Method;
