+++ server/src/main/java/io/spine/server/event/model/EventReactorMethod.java
@@ -23,7 +23,7 @@ package io.spine.server.event.model;
 import io.spine.server.event.EventReactor;
 import io.spine.server.event.React;
 import io.spine.server.model.EventProducingMethod;
-import io.spine.server.model.declare.ParameterSpec;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.type.EventClass;
 import io.spine.server.type.EventEnvelope;
