+++ server/src/main/java/io/spine/server/event/model/EventSubscriberMethod.java
@@ -26,7 +26,7 @@ import io.spine.base.FieldPaths;
 import io.spine.core.ByField;
 import io.spine.core.Subscribe;
 import io.spine.server.model.MessageFilter;
-import io.spine.server.model.declare.ParameterSpec;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.type.EventEnvelope;
 
 import java.lang.reflect.Method;
