+++ server/src/main/java/io/spine/server/event/model/StateSubscriberMethod.java
@@ -33,7 +33,7 @@ import io.spine.core.Subscribe;
 import io.spine.logging.Logging;
 import io.spine.server.model.MessageFilter;
 import io.spine.server.model.Model;
-import io.spine.server.model.declare.ParameterSpec;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.type.EventEnvelope;
 import io.spine.system.server.event.EntityStateChanged;
 import io.spine.type.TypeUrl;
