+++ server/src/main/java/io/spine/server/event/model/StateSubscriberSpec.java
@@ -26,14 +26,14 @@ import com.google.protobuf.Message;
 import io.spine.base.EventMessage;
 import io.spine.core.EventContext;
 import io.spine.server.entity.EntityVisibility;
-import io.spine.server.model.declare.ParameterSpec;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.type.EventEnvelope;
 import io.spine.system.server.event.EntityStateChanged;
 import io.spine.type.TypeName;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static io.spine.protobuf.AnyPacker.unpack;
-import static io.spine.server.model.declare.MethodParams.consistsOfTypes;
+import static io.spine.server.model.MethodParams.consistsOfTypes;
 
 /**
  * A {@link ParameterSpec} of an entity state subscriber method.
