+++ server/src/main/java/io/spine/server/event/model/SubscriberMethod.java
@@ -27,8 +27,8 @@ import io.spine.server.event.EventSubscriber;
 import io.spine.server.model.AbstractHandlerMethod;
 import io.spine.server.model.HandlerId;
 import io.spine.server.model.MessageFilter;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.model.VoidMethod;
-import io.spine.server.model.declare.ParameterSpec;
 import io.spine.server.type.EmptyClass;
 import io.spine.server.type.EventClass;
 import io.spine.server.type.EventEnvelope;
