+++ server/src/main/java/io/spine/server/model/AccessModifier.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.model.declare;
+package io.spine.server.model;
 
 import com.google.common.base.Joiner;
 import com.google.common.collect.ImmutableList;
