+++ server/src/main/java/io/spine/server/model/ParameterSpec.java
@@ -18,10 +18,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.model.declare;
+package io.spine.server.model;
 
 import com.google.errorprone.annotations.Immutable;
-import io.spine.server.model.HandlerMethod;
 import io.spine.server.type.MessageEnvelope;
 
 /**
