+++ server/src/test/java/io/spine/server/aggregate/model/ApplierTest.java
@@ -32,8 +32,8 @@ import io.spine.server.aggregate.Apply;
 import io.spine.server.aggregate.model.EventApplierSignature.EventApplierParams;
 import io.spine.server.dispatch.Success;
 import io.spine.server.model.IllegalOutcomeException;
-import io.spine.server.model.declare.MatchCriterion;
-import io.spine.server.model.declare.SignatureMismatch;
+import io.spine.server.model.MatchCriterion;
+import io.spine.server.model.SignatureMismatch;
 import io.spine.server.test.shared.EmptyAggregate;
 import io.spine.server.type.EventEnvelope;
 import io.spine.test.reflect.event.RefProjectCreated;
