+++ server/src/test/java/io/spine/server/event/model/EventSubscriberMethodTest.java
@@ -36,7 +36,7 @@ import io.spine.server.event.model.given.subscriber.InvalidTwoParamsSecondInvali
 import io.spine.server.event.model.given.subscriber.TestEventSubscriber;
 import io.spine.server.event.model.given.subscriber.ValidButPrivate;
 import io.spine.server.event.model.given.subscriber.ValidOneParam;
-import io.spine.server.model.declare.SignatureMismatchException;
+import io.spine.server.model.SignatureMismatchException;
 import io.spine.server.model.given.Given;
 import io.spine.server.type.EventEnvelope;
 import io.spine.test.reflect.event.RefProjectCreated;
