+++ server/src/test/java/io/spine/server/model/given/method/OneParamSpec.java
@@ -22,10 +22,10 @@ package io.spine.server.model.given.method;
 
 import com.google.errorprone.annotations.Immutable;
 import io.spine.base.EventMessage;
-import io.spine.server.model.declare.ParameterSpec;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.type.EventEnvelope;
 
-import static io.spine.server.model.declare.MethodParams.consistsOfSingle;
+import static io.spine.server.model.MethodParams.consistsOfSingle;
 
 @Immutable
 public enum OneParamSpec implements ParameterSpec<EventEnvelope> {
