+++ server/src/test/java/io/spine/server/model/given/method/TwoParamMethod.java
@@ -23,8 +23,8 @@ package io.spine.server.model.given.method;
 import io.spine.base.EventMessage;
 import io.spine.server.model.AbstractHandlerMethod;
 import io.spine.server.model.HandlerId;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.model.VoidMethod;
-import io.spine.server.model.declare.ParameterSpec;
 import io.spine.server.type.EmptyClass;
 import io.spine.server.type.EventClass;
 import io.spine.server.type.EventEnvelope;
