+++ server/src/test/java/io/spine/server/model/given/method/TwoParamSpec.java
@@ -23,10 +23,10 @@ package io.spine.server.model.given.method;
 import com.google.errorprone.annotations.Immutable;
 import com.google.protobuf.Message;
 import io.spine.core.EventContext;
-import io.spine.server.model.declare.ParameterSpec;
+import io.spine.server.model.ParameterSpec;
 import io.spine.server.type.EventEnvelope;
 
-import static io.spine.server.model.declare.MethodParams.consistsOfTwo;
+import static io.spine.server.model.MethodParams.consistsOfTwo;
 
 @Immutable
 public enum TwoParamSpec implements ParameterSpec<EventEnvelope> {
